[cuda][record-only] fp8 e5m2 KV cache attempt (DO-NOT-SHIP: decode flat 49.03, does not cross 50)
Record-only branch off g4-opt-gs64 (5191475c8b). NOT in the shipped chain, NOT the tip.
fp8 e5m2 KV: decode 49.03 (flat vs tip 49.14; fresh tip 48.89) - does not cross 50;
prefill 1617 (+35.5%, off-target); token-agreement 77.8% (e5m2 2-mantissa-bit lossy);
.ptd KV halved (22.64->21.63GB). nsys: decode SDPA split-K only -3.4% (not ~40%), ~6.6%
of decode -> weight-BW bound (int4/int8 matvecs unchanged). See RESULTS_g4-opt-fp8kv.md.

Export-side: model.py (KV bf16->e5m2), export.py, sdpa.py (_validate_sdpa_inputs relax),
replacement_pass.py. Runner shims (additive, backward-compatible): common_shims_slim.cpp/.h
+ slim ScalarType.h (Float8_e5m2=23 enum+elementSize+isValid+toString+isFloatingType) +
common_shims.cpp/.h + utils.h (ATen/Metal path).

+++ examples/models/gemma4/text_decoder/gemma4_attention.py
@@ -116,8 +116,8 @@ class Gemma4KVCache(nn.Module):
             Tuple of (full_k, full_v) - returns entire cache
         """
         if self.use_index_copy:
-            self.k_cache.index_copy_(2, input_pos, k_val)
-            self.v_cache.index_copy_(2, input_pos, v_val)
+            self.k_cache.index_copy_(2, input_pos, k_val.to(self.k_cache.dtype))
+            self.v_cache.index_copy_(2, input_pos, v_val.to(self.v_cache.dtype))
         else:
             seq_len = k_val.size(2)
             start_pos = input_pos[0].item()
